移動 / CLIコマンドの実行部分を新しいクラスに分離
CLIコマンドを MainWindowViewModel 以外でも利用できるようにするため
